fix(file): keymap improvement
